Require 'static for S to prevent unsound behaviour
If a reference is used for S (such as `&mut X`) then something as
simple as the following results in a use-after-free:

pub async fn foo(handle: Handle<&mut X, ()>) {
    let mut x = X::new();
    handle.yield_value(&mut x).await;
}

Requiring `S` to be 'static prevents this from compiling and avoids
the issue but it would be nice to find a way to tie the lifetime of
`S` to the parent `Coro` if possible. I've not found a way to do this
yet and I suspect that it isn't currently possible at all, but I may
be wrong on that front.